requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
idna==3.10
pillow==11.2.1
python-dotenv==1.1.0
qrcode==8.2
colorama==0.4.6
lxml==6.0.1
python-pptx==1.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
xlsxwriter==3.2.9
